Moved `write_funcs_to_file` from hyperopt_utils.
This function makes more sense in this repository; though it's useful for working with hyperopt, it's functionality really just has to do with writing a file that contains the code to run functions defined in a notebook. (haven't tested using the function after moving it here, but I don't think any imports, etc. are missing)